+++ CHANGELOG.md
@@ -7,24 +7,36 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Add Category/Brand name in `titleTag` when meta tag title are null.
+
 ## [2.71.2] - 2019-05-07
+
 ### Fixed
+
 - Slugify facets when checking if is selected.
 
 ## [2.71.1] - 2019-05-07
 
 ## [2.71.0] - 2019-05-07
+
 ### Added
+
 - Added `href` field to facets category. This field should be used to populate links in the store
 
 ### Changed
+
 - SEGMENT scope to root field resolvers containing translatable fields
 
 ### Fixed
+
 - Uses default sales channel as API language
 
 ## [2.70.4] - 2019-05-06
+
 ### Added
+
 - Parameter `hideUnavailableItems` to catalog search queries.
 
 ## [2.70.3] - 2019-05-06